Modified the load genes endpoint to accept a file path
