+++ src/chanjo2/constants.py
@@ -13,7 +13,6 @@ MULTIPLE_GENE_LISTS_NOT_SUPPORTED_MSG = (
     "Please provide either Ensembl gene IDs, HGNC gene IDS or HGNC gene symbols."
 )
 AMBIGUOUS_SAMPLES_INPUT = "Please provide either a name of a case or a list of samples."
-MULTIPLE_GENOMIC_INTERVAL_PARAMS_NOT_SUPPORTED = "Either genome build or the path to a local file containing the resources should be provided, not both."
 
 ENSEMBL_RESOURCE_CLIENT: Dict[str, Callable] = {
     "genes": fetch_ensembl_genes,
